[Performance] Remove redundant float32 cast in DeepSeek V4 router logits
The custom NPU kernels (moe_gating_top_k / moe_gating_top_k_hash)
already handle bfloat16 input and internally cast to float32 for
softmax/top-k compute. The Python-level .float() cast wasted memory
bandwidth and forced the linear projection to run in float32 instead
of bfloat16 tensor cores.

Benchmark results on 8x910B3 with DeepSeek-V4-Flash-w8a8-mtp:
- QPS=1 mean TTFT: 936ms -> 653ms (-30%)
- QPS=1 p99 TTFT: 5286ms -> 1408ms (-73%)
- QPS=16 throughput: +3.6%
- QPS=inf throughput: +2.1%

+++ vllm_ascend/models/deepseek_v4.py
@@ -351,7 +351,7 @@ class DeepseekV4MoE(nn.Module):
                                          router_logits=hidden_states)
         else:
             # router_logits: (num_tokens, n_experts)
-            router_logits = F.linear(hidden_states.float(), self.gate.weight)
+            router_logits = F.linear(hidden_states, self.gate.weight)
             fused_moe_out = self.experts(hidden_states=hidden_states,
                                          router_logits=router_logits)
